x86/highmem: close race between clear/set ptes
If the task is interrupted during a kmap_atomic() / kunmap_atomic() (or
the same code in kmap_atomic_prot_pfn() and its counter part) it may
race against switch_kmaps() and trigger a false positive warning.
In kmap_atomic_prot() we first grab a new index via
kmap_atomic_idx_push() and then check if the slot is already in use.

If we get interrupted after taking the index then switch_kmaps() will
assume that the index is in use and write the old entry from the
kmap_pte member. Since __kunmap_atomic() never invalidates this member
it might write an old entry and now it looks like this entry is already
in use and a WARN_ON() is seen.
This patch sets the shadow pte entry to 0 so pte_none() doesn't trigger
a warning.
While here, I add a BUG_ON() to kmap_atomic_idx_push() which is also in
the non-RT case.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>