net/cpsw: use a lock around source testing
For some reason on RT it happens that the TX interrupt fires over and
over again but according to tx_stat there is nothing going on. Same goes
for RX but not that often.
With this lock around it this is gone. However I still see from time to
time interrupts where each source is set to 0.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>